+++ app/Http/Controllers/Backend/RoleController.php
@@ -8,6 +8,7 @@ use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Brian2694\Toastr\Facades\Toastr;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\RoleStoreRequest;
 use App\Http\Requests\RoleUpdateRequest;
 
@@ -18,6 +19,7 @@ class RoleController extends Controller
      */
     public function index()
     {
+        Gate::authorize('index-role'); // authorize this user to access/give access to admin dashboard
         $roles = Role::with(['permissions:id,permission_name,permission_slug'])->select(['id', 'role_name', 'is_deleteable', 'updated_at'])->get();
         return view('admin.pages.role.index', compact('roles'));
     }
@@ -27,6 +29,7 @@ class RoleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create-role'); // authorize this user to access/give access to admin dashboard
         $modules = Module::with(['permissions:id,module_id,permission_name,permission_slug'])->select(['id', 'module_name'])->get();
         // return $modules;
         return view('admin.pages.role.create', compact('modules'));
@@ -37,6 +40,7 @@ class RoleController extends Controller
      */
     public function store(RoleStoreRequest $request)
     {
+        Gate::authorize('create-role'); // authorize this user to access/give access to admin dashboard
         Role::updateOrCreate([
             'role_name' => $request->role_name,
             'role_slug' => Str::slug($request->role_name),
@@ -60,6 +64,7 @@ class RoleController extends Controller
      */
     public function edit($id)
     {
+        Gate::authorize('edit-role'); // authorize this user to access/give access to admin dashboard
         $role = Role::with(['permissions'])->find($id);
         $modules = Module::with(['permissions:id,module_id,permission_name,permission_slug'])->select(['id', 'module_name'])->get();
         return view('admin.pages.role.edit', compact('modules', 'role'));
@@ -70,6 +75,7 @@ class RoleController extends Controller
      */
     public function update(RoleUpdateRequest $request, $id) //(Request $request, $id)
     {
+        Gate::authorize('edit-role'); // authorize this user to access/give access to admin dashboard
         // dd($id);
         $role = Role::find($id);
         $role->update([
@@ -88,6 +94,7 @@ class RoleController extends Controller
      */
     public function destroy($id)
     {
+        Gate::authorize('delete-role'); // authorize this user to access/give access to admin dashboard
         $role = Role::find($id);
         if($role->is_deleteable){
             $role->delete();
